api_scripts/cloud_functions/storage_bigquery/requirements.txt
# python packages dependencies 
pandas==1.3.5
google-cloud-storage==1.30.0
google-cloud-bigquery

